fix for seeing purchaseable items without license metadata
i noticed if i didnt have a weapons license granted but had the item i was still able to see and purchase weapons. added 
```
and hasLicense
``` 
to the canSee global to check if required

tested by granting license and having the item. was able to see. revoked license, couldnt see :)